pyproject.toml
[tool.poetry]
name = "whisper-segmentation"
version = "0.1.0"
description = ""
authors = ["Ryu1845 <ryu@tpgjbo.xyz>"]
license = "GPLv3"

[tool.poetry.dependencies]
python = ">=3.8.1,<4.0"
ffmpeg-python = "^0.2.0"
typed-argument-parser = "^1.7.2"
pywhisper = "^1.0.6"

[tool.poetry.dev-dependencies]
black = "^22.10.0"
isort = "^5.10.1"
ssort = "^0.11.6"
flake8 = "^6.0.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
